requirements.txt
pynput
sounddevice
soundfile
faster-whisper
pystray
Pillow